requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.10
backcall==0.2.0
beautifulsoup4==4.6.3
billiard==3.3.0.23
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
colorlog==4.1.0
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
decorator==4.4.2
Django==2.1.3
django-cors-headers==2.4.0
django-filter==2.3.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
drf-extensions==0.6.0
drf-yasg==1.17.0
emoji==0.6.0
et-xmlfile==1.1.0
Faker==2.0.1
har2case==0.3.1
idna==2.8
inflection==0.3.1
ipython==7.16.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.17.2
Jinja2==2.10.3
kombu==3.0.37
MarkupSafe==1.1.1
mysqlclient==1.4.4
Naked==0.1.31
openpyxl==3.0.7
packaging==19.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.6
ptyprocess==0.7.0
pycryptodome==3.9.9
Pygments==2.6.1
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
shellescape==3.8.1
six==1.13.0
sqlparse==0.3.1
text-unidecode==1.2
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.7
uWSGI==2.0.19.1
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
xmindparser==1.0.8
apscheduler
